finalisation du projet

+++ controllers/sauces.js
@@ -37,12 +37,18 @@ exports.getOneSauce = (req, res, next) => {
 exports.modifySauce = (req, res, next) => {
     const sauceObject = req.file
         ? {
-              ...JSON.parse(req.body.sauce),
-              imageUrl: `${req.protocol}://${req.get("host")}/images/${
-                  req.file.filename
-              }`,
-          }
+            ...JSON.parse(req.body.sauce),
+            imageUrl: `${req.protocol}://${req.get("host")}/images/${
+                req.file.filename
+            }`,
+        }
         : { ...req.body };
+        //operateur ternaire(condition ? valeur si vrai : valeur si faux)
+        if (req.auth.userId !== sauceObject.userId) {
+            return res
+                .status(403)
+                .json({ error: "opération non autorisée" });
+        }
     Sauce.updateOne(
         { _id: req.params.id },
         { ...sauceObject, _id: req.params.id }
@@ -54,6 +60,11 @@ exports.modifySauce = (req, res, next) => {
 exports.deleteSauce = (req, res, next) => {
     Sauce.findOne({ _id: req.params.id })
         .then((sauce) => {
+            if (req.auth.userId !== sauce.userId) {
+                return res
+                    .status(403)
+                    .json({ error: "opération non autorisée" });
+            }
             const filename = sauce.imageUrl.split("/images/")[1];
             fs.unlink(`images/${filename}`, () => {
                 Sauce.deleteOne({ _id: req.params.id })
@@ -88,11 +99,11 @@ exports.like = (req, res, next) => {
                     users.push(req.body.userId);
                     Sauce.updateOne(
                         { _id: req.params.id },
-                        { likes: compteurLikes,usersLiked: users}
+                        { likes: compteurLikes, usersLiked: users }
                     )
                         .then(() => {
-                            console.log("liké ou non");
-                            res.status(200).json({ message: "liké !" });
+                            console.log("liké");
+                            res.status(200).json({ message: "liké ok !" });
                         })
 
                         .catch((error) => {
@@ -109,21 +120,69 @@ exports.like = (req, res, next) => {
                     users2.push(req.body.userId);
                     Sauce.updateOne(
                         { _id: req.params.id },
-                        { dislikes: compteurDislikes,usersDisliked:users2 }
+                        { dislikes: compteurDislikes, usersDisliked: users2 }
                     )
                         .then(() => {
-                            res.status(200).json({ message: "liké !" });
+                            console.log("disliké");
+                            res.status(200).json({ message: "disliké ok !" });
                         })
 
                         .catch((error) => {
-                          res.status(400).json({
+                            res.status(400).json({
                                 error: error,
                             });
                         });
+                    break;
                 default:
-                        if(sauce.usersLiked.includes(req.body.userId)){
-                          
-                        }
+                    if (sauce.usersLiked.includes(req.body.userId)) {
+                        let compteurLikes = sauce.likes || 0;
+                        compteurLikes--;
+                        let users = sauce.usersLiked || [];
+                        users = users.filter((uId) => uId != req.body.userId);
+                        Sauce.updateOne(
+                            { _id: req.params.id },
+                            { likes: compteurLikes, usersLiked: users }
+                        )
+                            .then(() => {
+                                console.log("retrait du like");
+                                res.status(200).json({
+                                    message: "retrait du like ok !",
+                                });
+                            })
+
+                            .catch((error) => {
+                                console.log(error);
+                                res.status(400).json({
+                                    error: error,
+                                });
+                            });
+                    } else {
+                        let compteurDislikes = sauce.dislikes || 0;
+                        compteurDislikes--;
+                        let users2 = sauce.usersDisliked || [];
+                        users2 = users2.filter(
+                            (u2Id) => u2Id != req.body.userId
+                        );
+                        Sauce.updateOne(
+                            { _id: req.params.id },
+                            {
+                                dislikes: compteurDislikes,
+                                usersDisliked: users2,
+                            }
+                        )
+                            .then(() => {
+                                console.log("retrait du dislike");
+                                res.status(200).json({
+                                    message: "retrait du dislike ok !",
+                                });
+                            })
+
+                            .catch((error) => {
+                                res.status(400).json({
+                                    error: error,
+                                });
+                            });
+                    }
                     break;
             }
         })
